[circuit] Separate library & elements modules
