+++ align/circuit/__init__.py
@@ -1,4 +1,4 @@
 from .core import Circuit, SubCircuit
-from .elements import Library
+from .library import Library
 from .parser import SpiceParser
 from .gen_dot import gen_dot_file
